+++ apps/teams/src/app/api/webhook/microsoft/event-handler/service.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, test, vi } from 'vitest';
 import { inngest } from '@/inngest/client';
-import type { WebhookPayload } from '@/app/api/microsoft/event-handler/service';
-import { EventType, handleWebhook } from '@/app/api/microsoft/event-handler/service';
-import type { SubscriptionPayload } from '@/app/api/microsoft/event-handler/types';
+import type { WebhookPayload } from '@/app/api/webhook/microsoft/event-handler/service';
+import { EventType, handleWebhook } from '@/app/api/webhook/microsoft/event-handler/service';
+import type { SubscriptionPayload } from '@/app/api/webhook/microsoft/event-handler/types';
 
 const channelData: SubscriptionPayload[] = [
   {
