+++ apps/teams/src/app/api/webhook/microsoft/event-handler/types.ts
@@ -1,5 +1,5 @@
 import type { z } from 'zod';
-import type { subscriptionSchema } from '@/app/api/microsoft/event-handler/schema';
+import type { subscriptionSchema } from '@/app/api/webhook/microsoft/event-handler/schema';
 
 export type SubscriptionPayload = z.infer<typeof subscriptionSchema>;
