+++ apps/teams/src/app/api/webhook/microsoft/lifecycle-notifications/schema.ts
