+++ apps/teams/src/app/api/webhook/microsoft/lifecycle-notifications/service.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test, vi } from 'vitest';
-import { handleSubscriptionEvent } from '@/app/api/microsoft/lifecycle-notifications/service';
+import { handleSubscriptionEvent } from '@/app/api/webhook/microsoft/lifecycle-notifications/service';
 import { inngest } from '@/inngest/client';
-import type { MicrosoftSubscriptionEvent } from '@/app/api/microsoft/lifecycle-notifications/types';
+import type { MicrosoftSubscriptionEvent } from '@/app/api/webhook/microsoft/lifecycle-notifications/types';
 
 const data: MicrosoftSubscriptionEvent[] = [
   {
