+++ apps/teams/src/app/api/webhook/microsoft/lifecycle-notifications/types.ts
@@ -1,4 +1,4 @@
 import type { z } from 'zod';
-import type { lifecycleEventSchema } from '@/app/api/microsoft/lifecycle-notifications/schema';
+import type { lifecycleEventSchema } from '@/app/api/webhook/microsoft/lifecycle-notifications/schema';
 
 export type MicrosoftSubscriptionEvent = z.infer<typeof lifecycleEventSchema>;
