+++ apps/teams/src/connectors/microsoft/subscriptions/subscriptions.test.ts
@@ -46,9 +46,9 @@ describe('subscriptions connector', () => {
           if (
             body.changeType !== subscriptionChangeType ||
             body.resource !== subscriptionResource ||
-            body.notificationUrl !== `${env.WEBHOOK_URL}/api/microsoft/event-handler` ||
+            body.notificationUrl !== `${env.WEBHOOK_URL}/api/webhook/microsoft/event-handler` ||
             body.lifecycleNotificationUrl !==
-              `${env.WEBHOOK_URL}/api/microsoft/lifecycle-notifications`
+              `${env.WEBHOOK_URL}/api/webhook/microsoft/lifecycle-notifications`
           ) {
             return new Response(undefined, { status: 400 });
           }
