+++ apps/teams/src/inngest/client.ts
@@ -1,7 +1,7 @@
 import { EventSchemas, type GetEvents, type GetFunctionInput, Inngest } from 'inngest';
 import { sentryMiddleware } from '@elba-security/inngest';
 import { logger } from '@elba-security/logger';
-import type { WebhookPayload } from '@/app/api/microsoft/event-handler/service';
+import type { WebhookPayload } from '@/app/api/webhook/microsoft/event-handler/service';
 import type { MessageMetadata } from '@/connectors/elba/data-protection/metadata';
 import { rateLimitMiddleware } from './middlewares/rate-limit-middleware';
