+++ apps/teams/src/middleware.ts
@@ -5,4 +5,4 @@ export const middleware = createElbaMiddleware({
   webhookSecret: env.ELBA_WEBHOOK_SECRET,
 });
 
-export const config = { matcher: ['/api/webhook/elba/(.*)', '/api/webhooks/elba/(.*)'] };
+export const config = { matcher: ['/api/webhooks/elba/(.*)'] };
